Paint the background of the legend white.
This corrects a usability problem with threadscope when used in an environment
with a theme where light text is used on a darker background, painting the
background white rather than allowing the default colour makes it possible to
view the legend in such an environment.  An example of this is when one uses
the "Hi Contrast Inverse" themes with Gnome.