change the log base in histogram to 10 to have more readable scales
Also, increase the number of bars slightly (for the test data it looks OK,
but may be wrong for other data). In the result the x-axis does not seem
to be cut off so abruptly at the right end (probably because the scale
constant is not the same as the log base).