Delete a FIXME about the starting time
The point time 0 of ghc-events is the start of the program and the start
of the INIT phase of RTS. We would prefer the end of the INIT phase
and the start of the MUT phase as the start of the maximal interval.
However, there is no ghc-event event at that point and the two events
currently in the middle of the INIT phase (Startup and WallClockTime)
are not stable enough (Startup will be removed and WallClockTime
is semantically not linked to the INIT phase in any way). So it's best
to stick to the 0 point as the start of the maximal interval in TS.