Mute all errors from GC finite state machine assumption failures
unless assertions are turned on. This is to avoid any future
breakages due to wrong events order, as in #35 (and probably in #46).
For any events where the order is indeed wrong, the TS GC stats for
that particular GC are more ore less inaccurate. This probably doesn't
happen if the eventlogs are collected on an OS with a good resolution
of a time source (e.g., on modern Linux).