Remove dependency on gtk-mac-integration
gtk2hs/gtk-mac-integration has been unmaintained for nearly 3 years
now and it makes it harder to build ThreadScope on macOS.

We've been using it only to display the logo in the Dock so the
effect of this removal is minimal. We can add this feature back
when porting ThreadScope to gi-gtk.