Switch to GTK3 (#137)
We upgrade from the deprecated GTK2 to the still supported GTK3.

Most of the changes we had to make were described in the migration guide:
https://docs.gtk.org/gtk3/migrating-2to3.html#changes-that-need-to-be-done-at-the-time-of-the-switch

Many of the functions we depended on for low-level rendering were removed, so I've substituted mostly equivalent ones.

The way events are subscribed to has slightly changed and that has lead to a bunch of changes.

With the new version of GTK3, windows CI works again. Resolves #135, #97

We've also updated the Cabal file to be made aware of stuff included through TH.